Move IgnoredMeshList to Towers.
